feat: sale event 관련 함수 분리
